refactor(`JekyllSiteProvider`, `CategoryItem`): seperate `CategoryItem` from `JekyllSiteProvider`
